docs: introduce OSLinkTxQueueUsed (ROM call 0x4F4)
OSLinkTxQueueUsed is the inverse of OsLinkTxQueueInquire: rather than
returning the amount of free space in the tx buffer, it returns the
space used.

The routine is only used by AMS as a boolean in OSLinkOpen to determine
if link flushing is needed. I cannot find any other use of it in AMS,
nor think of any other possible use not covered by OSLinkTxQueueInquire.